Rename the geos_area.py with a leading underscore, to indicate that it contains private functions not designed for the user to use independently of satpy readers
