+++ satpy/readers/_geos_area.py
