+++ satpy/readers/agri_l1.py
@@ -28,7 +28,7 @@ import numpy as np
 import xarray as xr
 import dask.array as da
 from datetime import datetime
-from satpy.readers.geos_area import get_area_extent, get_area_definition
+from satpy.readers._geos_area import get_area_extent, get_area_definition
 from satpy.readers.hdf5_utils import HDF5FileHandler
 
 logger = logging.getLogger(__name__)
