+++ satpy/readers/electrol_hrit.py
@@ -51,7 +51,7 @@ from datetime import datetime
 import numpy as np
 import xarray as xr
 
-from satpy.readers.geos_area import get_area_extent, get_area_definition
+from satpy.readers._geos_area import get_area_extent, get_area_definition
 from satpy.readers.hrit_base import (HRITFileHandler, ancillary_text,
                                      annotation_header, base_hdr_map,
                                      image_data_function, time_cds_short)
